clean hqq_utils imports in hqq.py
